--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -36,23 +36,23 @@ jobs:
         os: [ubuntu-latest]
         python-version: ["3.8", "3.9", "3.10", "3.11", "3.12", "pypy-3.9"]
         optional-deps: [true]
-        with-python-isal: [true]
+        with-libs: [true]
         include:
         - os: macos-latest
           python-version: "3.10"
           optional-deps: true
         - os: ubuntu-20.04
           python-version: "3.10"
-          with-python-isal: false
+          with-libs: false
           optional-deps: false
         - os: ubuntu-20.04
           python-version: "3.10"
-          with-python-isal: false
+          with-libs: false
           optional-deps: true
         - os: ubuntu-20.04
           python-version: "3.10"
           optional-deps: false
-          with-python-isal: false
+          with-libs: false
           with-zstandard: true
         - os: windows-latest
           python-version: "3.10"
@@ -77,10 +77,10 @@ jobs:
       run: python -m pip install tox
     - name: Test
       run: tox -e py
-      if: matrix.with-python-isal
-    - name: Test without python-isal
-      run: tox -e no-isal
-      if: true && !matrix.with-python-isal
+      if: matrix.with-libs
+    - name: Test without python-isal and python-zlib-ng
+      run: tox -e no-libs
+      if: true && !matrix.with-libs
     - name: Test with zstandard
       if: matrix.with-zstandard
       run: tox -e zstd