--- a/tox.ini
+++ b/tox.ini
@@ -21,9 +21,9 @@ deps =
     {[testenv]deps}
     zstandard
 
-[testenv:no-isal]
+[testenv:no-libs]
 commands=
-    pip uninstall -y isal
+    pip uninstall -y isal zlib-ng
     {[testenv]commands}
 
 [testenv:black]