@rhpvorderman 2024-01-09
Hi @marcelm, back from the holidays yet? This will have a nice performance impact on cutadapt with default settings. I guess the PR e-mail will be at the bottom of a long list of other e-mails by now, so this is just a friendly ping to bump it back into view.