@marcelm 2024-01-12
Hi, indeed, the GitHub notifications have piled up quite a bit ...

This looks good! Can you please add a chagelog entry?

I was considering changing the default compression level in Cutadapt from 5 to 4. Do I understand correctly that with this change, I would get a *smaller* file *and* that it would be  twice as fast? (At least for this dataset.)