@rhpvorderman 2024-01-12
> This looks good! Can you please add a chagelog entry?

Done. 

> I was considering changing the default compression level in Cutadapt from 5 to 4. Do I understand correctly that with this change, I would get a smaller file and that it would be twice as fast? (At least for this dataset.)

Yes and yes :-D. Though I still recommend level 2, that is almost an order of magnitude faster ;-).